Move ExprEvalRecord/ExprEvalScratch to new ExprEvalComponent
